refactor(move constants): create constants folder and use @ syntax for import
